requirements.txt
tensorflow
keras
gemini-python
nltk
re